Research Assistant Agent/pyproject.toml
[project]
name = "research-assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
